+++ .gitignore
@@ -1,6 +1,8 @@
 email-subscribers.json
 email-rate-limit.json
 
+version.txt
+
 *.db
 
 *.db
