+++ app/__init__.py
@@ -8,6 +8,7 @@ from flask_cors import CORS
 import atexit
 import logging
 from os import getenv
+from .version import __version__
 from datetime import datetime
 
 
@@ -59,6 +60,7 @@ from .bp.email_subscription import email_subscription_bp
 from .discord import discord_notifier
 from .config import Config
 from .database import db, init_db
+from .bp.healthcheck import bp_healthcheck
 
 app = Flask(__name__)
 app.config.from_object(Config)
@@ -74,7 +76,8 @@ CORS(app, resources = {
 })
 
 app.register_blueprint(email_subscription_bp, url_prefix='/api')
-
+app.register_blueprint(bp_healthcheck, url_prefix='/')
+logger.info("SAY Website Backend version %s starting up", __version__)
 # Send startup notification
 if not app.debug and getenv("FLASK_ENV") != "development":
     discord_notifier.send_startup_notification("SAY Website Backend")
@@ -127,193 +130,6 @@ def handle_internal_error(error):
     
     return "An internal server error occurred. Please try again later.", 500
 
-@app.route("/api/health")
-@app.route("/api/healthcheck")
-@app.route("/health")
-@app.route("/healthcheck")
-def health():
-    """Comprehensive health check for all system components"""
-    from datetime import datetime
-    from .models.email import EmailSubscriber, EmailRateLimit
-    import os
-    
-    health_status = {
-        "status": "healthy",
-        "timestamp": datetime.utcnow().isoformat(),
-        "checks": {},
-        "version": "1.0.0",
-        "environment": "development" if app.debug else "production"
-    }
-    
-    overall_healthy = True
-    
-    # Database connectivity check
-    try:
-        # Test database connection by doing a simple query
-        subscriber_count = EmailSubscriber.query.count()
-        rate_limit_count = EmailRateLimit.query.count()
-        
-        health_status["checks"]["database"] = {
-            "status": "healthy",
-            "message": "Database connection successful",
-            "details": {
-                "subscribers_count": subscriber_count,
-                "rate_limits_count": rate_limit_count,
-                "database_url": app.config.get('SQLALCHEMY_DATABASE_URI', '').split('://')[0] + "://***"
-            }
-        }
-    except Exception as e:
-        overall_healthy = False
-        health_status["checks"]["database"] = {
-            "status": "unhealthy",
-            "message": f"Database connection failed: {str(e)}",
-            "details": {}
-        }
-    
-    # Email configuration check
-    try:
-        smtp_password = os.getenv("GOOGLE_APP_PASSWORD")
-        email_address = os.getenv("EMAIL", "stanthonyyouth.noreply@gmail.com")
-        
-        if smtp_password:
-            health_status["checks"]["email_config"] = {
-                "status": "healthy",
-                "message": "Email configuration present",
-                "details": {
-                    "smtp_configured": True,
-                    "email_address": email_address,
-                    "actually_send_email": not os.getenv("NO_EMAIL")
-                }
-            }
-        else:
-            health_status["checks"]["email_config"] = {
-                "status": "warning",
-                "message": "SMTP password not configured",
-                "details": {
-                    "smtp_configured": False,
-                    "email_address": email_address,
-                    "actually_send_email": False
-                }
-            }
-    except Exception as e:
-        health_status["checks"]["email_config"] = {
-            "status": "unhealthy",
-            "message": f"Email configuration check failed: {str(e)}",
-            "details": {}
-        }
-    
-    # Environment variables check
-    try:
-        required_vars = ["DATABASE_URL", "EMAIL"]
-        optional_vars = ["GOOGLE_APP_PASSWORD", "NO_EMAIL", "EMAIL_RATE_LIMIT_PER_DAY"]
-        
-        env_status = {}
-        missing_required = []
-        
-        for var in required_vars:
-            if os.getenv(var):
-                env_status[var] = "present"
-            else:
-                env_status[var] = "missing"
-                missing_required.append(var)
-        
-        for var in optional_vars:
-            env_status[var] = "present" if os.getenv(var) else "missing"
-        
-        if missing_required:
-            health_status["checks"]["environment"] = {
-                "status": "warning",
-                "message": f"Missing required environment variables: {', '.join(missing_required)}",
-                "details": env_status
-            }
-        else:
-            health_status["checks"]["environment"] = {
-                "status": "healthy",
-                "message": "All required environment variables present",
-                "details": env_status
-            }
-    except Exception as e:
-        health_status["checks"]["environment"] = {
-            "status": "unhealthy",
-            "message": f"Environment check failed: {str(e)}",
-            "details": {}
-        }
-    
-    # Application configuration check
-    try:
-        config_status = {
-            "debug_mode": app.debug,
-            "testing": app.testing,
-            "rate_limit_per_day": app.config.get('EMAIL_RATE_LIMIT_PER_DAY', 2),
-            "secret_key_configured": bool(app.secret_key),
-            "sqlalchemy_track_modifications": app.config.get('SQLALCHEMY_TRACK_MODIFICATIONS', True)
-        }
-        
-        health_status["checks"]["application"] = {
-            "status": "healthy",
-            "message": "Application configuration loaded",
-            "details": config_status
-        }
-    except Exception as e:
-        overall_healthy = False
-        health_status["checks"]["application"] = {
-            "status": "unhealthy",
-            "message": f"Application configuration check failed: {str(e)}",
-            "details": {}
-        }
-    
-    # API endpoints check
-    try:
-        from flask import url_for
-        endpoints = []
-        for rule in app.url_map.iter_rules():
-            if rule.endpoint != 'static':
-                methods = rule.methods or set()
-                filtered_methods = list(methods - {'HEAD', 'OPTIONS'}) if methods else []
-                endpoints.append({
-                    "endpoint": rule.endpoint,
-                    "methods": filtered_methods,
-                    "rule": str(rule)
-                })
-        
-        health_status["checks"]["api_endpoints"] = {
-            "status": "healthy",
-            "message": f"Found {len(endpoints)} API endpoints",
-            "details": {
-                "endpoint_count": len(endpoints),
-                "endpoints": endpoints[:10]  # Limit to first 10 for brevity
-            }
-        }
-    except Exception as e:
-        health_status["checks"]["api_endpoints"] = {
-            "status": "unhealthy",
-            "message": f"API endpoints check failed: {str(e)}",
-            "details": {}
-        }
-    
-    # Set overall status
-    unhealthy_checks = [check for check in health_status["checks"].values() if check["status"] == "unhealthy"]
-    warning_checks = [check for check in health_status["checks"].values() if check["status"] == "warning"]
-    
-    if unhealthy_checks:
-        health_status["status"] = "unhealthy"
-        status_code = 503
-    elif warning_checks:
-        health_status["status"] = "degraded"
-        status_code = 200
-    else:
-        health_status["status"] = "healthy"
-        status_code = 200
-    
-    health_status["summary"] = {
-        "total_checks": len(health_status["checks"]),
-        "healthy_checks": len([c for c in health_status["checks"].values() if c["status"] == "healthy"]),
-        "warning_checks": len(warning_checks),
-        "unhealthy_checks": len(unhealthy_checks)
-    }
-    
-    return jsonify(health_status), status_code
-
 @app.errorhandler(404)
 def handle_not_found(error):
     """Handle 404 errors."""
@@ -324,28 +140,6 @@ def handle_not_found(error):
 def index():
     """Health check endpoint."""
     return "All systems operational. API is running."
-
-@app.route('/health')
-def health_check():
-    """Detailed health check endpoint."""
-    discord_healthy = discord_notifier.is_healthy()
-    queue_size = discord_notifier.get_queue_size()
-    rate_limit_info = discord_notifier.get_rate_limit_info()
-    
-    health_status = {
-        "status": "healthy",
-        "timestamp": datetime.now().isoformat(),
-        "services": {
-            "discord_notifications": {
-                "status": "healthy" if discord_healthy else "unhealthy",
-                "enabled": discord_notifier.enabled,
-                "queue_size": queue_size,
-                "rate_limits": rate_limit_info
-            }
-        }
-    }
-    
-    return jsonify(health_status)
 
 # Register shutdown handler to gracefully close Discord notification system
 @atexit.register
