+++ app/version.py
@@ -1,0 +1,19 @@
+# Copyright (c) 2025 Damien Boisvert (AlphaGameDeveloper)
+# 
+# This software is released under the MIT License.
+# https://opensource.org/licenses/MIT
+from os.path import isfile
+
+# File for tracking the application version
+
+
+try:
+    from setuptools_scm import get_version
+    __version__ = get_version()
+
+except (ImportError, LookupError, OSError):
+    try:
+        with open("version.txt", "r") as f:
+            __version__ = f.read().strip()
+    except (FileNotFoundError, IOError):
+        __version__ = "0.0.0"
